Added checkout component for Stripe functionality
Allows for Checkout button to appear for each image in AllImages
Removed old checkout feature
