+++ src/components/App/App.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react'
 import { Route } from 'react-router-dom'
-import { CardElement, Elements, useElements, useStripe } from '@stripe/react-stripe-js'
-import { loadStripe } from '@stripe/stripe-js'
+// import { CardElement, Elements, useElements, useStripe } from '@stripe/react-stripe-js'
+// import { loadStripe } from '@stripe/stripe-js'
 // import image from 'Image.js'
 
 import AuthenticatedRoute from '../AuthenticatedRoute/AuthenticatedRoute'
@@ -31,7 +31,6 @@ import CommentDelete from '../routes/CommentDelete'
 import HomePage from '../routes/HomePage'
 import UploadS3Image from '../routes/ImageCreate'
 import ImageTag from '../routes/ImageTag'
-import ProductDisplay from './Checkout'
 // import Application from './Checkout.js'
 
 // const Message = ({ message }) => (
@@ -40,64 +39,64 @@ import ProductDisplay from './Checkout'
 //   </section>
 // )
 
-const CheckoutForm = () => {
-  const stripe = useStripe()
-  const elements = useElements()
-
-  const handleSubmit = async (event) => {
-    // Block native form submission
-    event.preventDefault()
-
-    if (!stripe || !elements) {
-      // Stripe.js has not loaded yet. Make sure to disable
-      // form submission until Stripe.js has loaded
-      return
-    }
-
-    // Get a reference to a mounted CardElement.
-    const cardElement = elements.getElement(CardElement)
-
-    // Use your card Element with other Stripe.js APIs
-    const { error, paymentMethod } = await stripe.createPaymentMethod({
-      type: 'card',
-      card: cardElement
-    })
-
-    if (error) {
-      console.log('[error], error')
-    } else {
-      console.log('[PaymentMethod]', paymentMethod)
-    }
-  }
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <CardElement
-        options={{
-          style: {
-            base: {
-              fontSize: '16px',
-              color: '#424770',
-              '::placeholder': {
-                color: '#aab7c4'
-              }
-            },
-            invalid: {
-              color: '#9e2146'
-            }
-          }
-        }}
-      />
-      <button type="submit" disable={!stripe}>
-          Pay
-      </button>
-    </form>
-  )
-}
+// const CheckoutForm = () => {
+//   const stripe = useStripe()
+//   const elements = useElements()
+//
+//   const handleSubmit = async (event) => {
+//     // Block native form submission
+//     event.preventDefault()
+//
+//     if (!stripe || !elements) {
+//       // Stripe.js has not loaded yet. Make sure to disable
+//       // form submission until Stripe.js has loaded
+//       return
+//     }
+//
+//     // Get a reference to a mounted CardElement.
+//     const cardElement = elements.getElement(CardElement)
+//
+//     // Use your card Element with other Stripe.js APIs
+//     const { error, paymentMethod } = await stripe.createPaymentMethod({
+//       type: 'card',
+//       card: cardElement
+//     })
+//
+//     if (error) {
+//       console.log('[error], error')
+//     } else {
+//       console.log('[PaymentMethod]', paymentMethod)
+//     }
+//   }
+//
+//   return (
+//     <form onSubmit={handleSubmit}>
+//       <CardElement
+//         options={{
+//           style: {
+//             base: {
+//               fontSize: '16px',
+//               color: '#424770',
+//               '::placeholder': {
+//                 color: '#aab7c4'
+//               }
+//             },
+//             invalid: {
+//               color: '#9e2146'
+//             }
+//           }
+//         }}
+//       />
+//       <button type="submit" disable={!stripe}>
+//           Pay
+//       </button>
+//     </form>
+//   )
+// }
 
 // Call 'loadStripe' outside of the component's render to avoid recreating 'Stripe'
 // object on every render
-const stripePromise = loadStripe('pk_test_51HobYFEybVIVldfc4QmD3NhroakMWJARBgzjLHf5tKx76TBTEmdcgnHrNFGujESH43KIdVM8xDur1JSCtaHqkQan00qUaWN889')
+// const stripePromise = loadStripe('pk_test_51HobYFEybVIVldfc4QmD3NhroakMWJARBgzjLHf5tKx76TBTEmdcgnHrNFGujESH43KIdVM8xDur1JSCtaHqkQan00qUaWN889')
 
 // const ProductDisplay = ({ handleClick }) => (
 //   <section>
@@ -191,11 +190,6 @@ class App extends Component {
           <AuthenticatedRoute user={user} exact path='/all-images' render={(props) => (
             <AllImages {...props} msgAlert={this.msgAlert} user={user} />
           )} />
-          <Elements stripe={stripePromise} user={user}>
-            <CheckoutForm />
-          </Elements>
-          <ProductDisplay>
-          </ProductDisplay>
           <AuthenticatedRoute user={user} exact path='/create-blog' render={() => (
             <BlogCreate msgAlert={this.msgAlert} user={user}/>
           )} />
